# docs/api/openapi.yaml
{
  "components": {
    "schemas": {
      "AuditEvent": {
        "description": "State-changing action recorded by Corporate Control Tower.",
        "properties": {
          "action": {
            "examples": [
              "project.registered"
            ],
            "minLength": 3,
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "examples": [
              "system"
            ],
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "entity_id": {
            "examples": [
              "PSZ-2026"
            ],
            "minLength": 1,
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "examples": [
              "project"
            ],
            "minLength": 3,
            "title": "Entity Type",
            "type": "string"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          }
        },
        "required": [
          "actor",
          "action",
          "entity_type",
          "entity_id"
        ],
        "title": "AuditEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Portfolio governance and WEB SIG provisioning API for BIMSIG Enterprise.",
    "title": "Corporate Control Tower REV12",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit/events": {
      "get": {
        "description": "List recent audit events.",
        "operationId": "list_audit_events_api_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response List Audit Events Api V1 Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events"
      }
    },
    "/api/v1/companies": {
      "get": {
        "description": "List enterprise companies.",
        "operationId": "list_companies_api_v1_companies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  },
                  "title": "Response List Companies Api V1 Companies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Companies"
      },
      "post": {
        "description": "Register an enterprise company.",
        "operationId": "register_company_api_v1_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Company"
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "description": "Return one enterprise company.",
        "operationId": "get_company_api_v1_companies__company_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Company"
      }
    },
    "/api/v1/companies/{company_id}/dashboard/executive": {
      "get": {
        "description": "Return the company-scoped executive dashboard read model.",
        "operationId": "executive_dashboard_api_v1_companies__company_id__dashboard_executive_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorporateDashboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Executive Dashboard"
      }
    },
    "/api/v1/companies/{company_id}/licenses": {
      "get": {
        "description": "List licenses assigned to a company.",
        "operationId": "list_company_licenses_api_v1_companies__company_id__licenses_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanyLicense"
                  },
                  "title": "Response List Company Licenses Api V1 Companies  Company Id  Licenses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Company Licenses"
      },
      "post": {
        "description": "Assign a license plan to a company.",
        "operationId": "assign_company_license_api_v1_companies__company_id__licenses_post",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyLicense"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyLicense"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Company License"
      }
    },
    "/api/v1/companies/{company_id}/memberships": {
      "get": {
        "description": "List memberships for a company.",
        "operationId": "list_company_memberships_api_v1_companies__company_id__memberships_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanyMembership"
                  },
                  "title": "Response List Company Memberships Api V1 Companies  Company Id  Memberships Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Company Memberships"
      },
      "post": {
        "description": "Assign a user role inside a company.",
        "operationId": "add_company_membership_api_v1_companies__company_id__memberships_post",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyMembership"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyMembership"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Company Membership"
      }
    },
    "/api/v1/companies/{company_id}/portfolio/summary": {
      "get": {
        "description": "Return portfolio status counts for one company.",
        "operationId": "company_portfolio_summary_api_v1_companies__company_id__portfolio_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Company Portfolio Summary Api V1 Companies  Company Id  Portfolio Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Portfolio Summary"
      }
    },
    "/api/v1/companies/{company_id}/projects": {
      "get": {
        "description": "List projects registered for one company.",
        "operationId": "list_company_projects_api_v1_companies__company_id__projects_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioProject"
                  },
                  "title": "Response List Company Projects Api V1 Companies  Company Id  Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Company Projects"
      },
      "post": {
        "description": "Register a project inside one company.",
        "operationId": "register_company_project_api_v1_companies__company_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioProject"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Company Project"
      }
    },
    "/api/v1/companies/{company_id}/projects/{project_id}": {
      "get": {
        "description": "Return one project inside one company.",
        "operationId": "get_company_project_api_v1_companies__company_id__projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioProject"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Company Project"
      }
    },
...
  }
}